refactor models and their using statements
