+++ src/BaseResolver.cs
@@ -4,6 +4,7 @@
 using System;
 
 using Microsoft.AspNetCore.Http;
+using Wangkanai.Detection.Models;
 using Wangkanai.Detection.Services;
 
 namespace Wangkanai.Detection
