+++ src/Models/Feature.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
-namespace Wangkanai.Detection
+namespace Wangkanai.Detection.Models
 {
     public class Feature
     {
