+++ src/Platform/PlatformFactory.cs
@@ -2,6 +2,7 @@
 // The Apache v2. See License.txt in the project root for license information.
 
 using System;
+using Wangkanai.Detection.Models;
 
 namespace Wangkanai.Detection
 {
@@ -9,7 +10,7 @@ namespace Wangkanai.Detection
     {
         public string? Name { get; set; }
         public string? Maker { get; set; }
-        public OperatingSystem OS { get; set; }
+        public Models.OperatingSystem OS { get; set; }
         public Processor CPU { get; set; }
         public Version? Version { get; set; }
