+++ src/Responsive/DeviceHttpContextExtensions.cs
@@ -4,6 +4,7 @@
 using System;
 
 using Wangkanai.Detection;
+using Wangkanai.Detection.Models;
 
 namespace Microsoft.AspNetCore.Http
 {
